+++ dciCommon.h
@@ -124,7 +124,7 @@ typedef enum dciType
 
 uint32_t dci0_lengthOfRIVviaBandwidth(bandwidth_t bandwidth);
 void dci0_CorrectnessParameters(uint8_t* dciParam, const uint8_t dci0_bandwidthPRB);
-void dci1_CorrectnessParameters(uint8_t* dciParam);
+void dci1_CorrectnessParameters(uint32_t* dciParam);
 uint32_t dci1_lengthOfBitmapViaBandwidth(bandwidth_t bandwidth);
 void dci60a_CorrectnessParameters(uint8_t* dciParam, const uint8_t dci60a_bandwidthPRB);
